One more pointer stored into a 32 bits int remove in mvl_parse.

+++ alliance/src/mbk/src/mvl_utype.h
@@ -28,5 +28,5 @@ struct dct_recrd
   short             fd4_val;
   short             fd5_val;
   short             fd6_val;
-  int               pnt_val;
+  long              pnt_val;
   };
